This Commit refrences updates to the docker files.
docker-compose: updted the reference name for the old server directory. it's now been renamed client_api to convey which service it provides. Added services for inital gmai syncing, gmail content fetching, and the placeholder for the relevance model. This is the exeutable that defines our services.

Dockerfile.client, Dockerfile.client_api, Dockerfile.worker -> define how each of the services should build the containers for each service.