updates login flow.
When a user logs in, we check the db to see if the user has been added to supa base. If they haven't we add them to the db.

The gmail linking process now happens inside of the user account page.

When a user clicks link account, oauth permission fires, and once granted begins the email extraction process.

If the user clicks unlink gmail, permisson is revoked, and the UI is updated to reflect that gmail is no longer connected.