Implemented front end fast api calls to the database for job application data.
This allows the client to fetch new emails from the database when the user lands on the home page. The data is then transformed into JobCards, and the rest of the logic flows from there. (Presenting the cards in the proper columns with template placeholders if data hasn't been loaded yet.)