Realtime event streaming from the database has been integrated.
After a user logs in, any changes to rows within the database for the user are broadcasted out, and the UI responds in realtime to said changes.

When new rows are entered, the UI adds a new job card. When existing rows are updated, the Job Cards are updated on the UI, and when rows are deleted, the job card is deleted from the UI.

This update keeps the UI and the database in sync while the user is logged in.